Prod

+++ src/main/java/com/nce/backend/cars/domain/valueObjects/ApiCarData.java
@@ -14,7 +14,7 @@ public record ApiCarData(
 
          String engineType,
 
-         Integer engineVolume,
+         Double engineVolume,
 
          String bodywork,
